requirements.txt
Flask==2.3.2
gunicorn==21.2.0
mysql-connector-python==8.0.33
matplotlib==3.7.1
numpy==1.26.5
